HSEARCH-3860 Move type-checking and type-conversion tests for field sorts to a dedicated class
Because we won't need them when testing sorts on multi-valued fields.
