+++ integrationtest/backend/tck/src/main/java/org/hibernate/search/integrationtest/backend/tck/search/sort/FieldSearchSortBaseIT.java
@@ -26,28 +26,20 @@ import org.hibernate.search.engine.backend.document.model.dsl.IndexSchemaElement
 import org.hibernate.search.engine.backend.document.model.dsl.IndexSchemaObjectField;
 import org.hibernate.search.engine.backend.document.model.dsl.ObjectFieldStorage;
 import org.hibernate.search.engine.backend.types.Sortable;
-import org.hibernate.search.engine.backend.types.dsl.IndexFieldTypeFactory;
 import org.hibernate.search.engine.backend.types.dsl.StandardIndexFieldTypeOptionsStep;
 import org.hibernate.search.engine.backend.work.execution.spi.IndexIndexingPlan;
-import org.hibernate.search.engine.reporting.spi.EventContexts;
-import org.hibernate.search.engine.search.common.ValueConvert;
 import org.hibernate.search.engine.search.query.SearchQuery;
 import org.hibernate.search.engine.search.sort.dsl.SearchSortFactory;
 import org.hibernate.search.engine.search.sort.dsl.SortFinalStep;
 import org.hibernate.search.integrationtest.backend.tck.testsupport.types.FieldTypeDescriptor;
 import org.hibernate.search.integrationtest.backend.tck.testsupport.types.NormalizedStringFieldTypeDescriptor;
 import org.hibernate.search.integrationtest.backend.tck.testsupport.util.ExpectationsAlternative;
-import org.hibernate.search.integrationtest.backend.tck.testsupport.util.InvalidType;
+import org.hibernate.search.integrationtest.backend.tck.testsupport.util.IndexFieldStructure;
 import org.hibernate.search.integrationtest.backend.tck.testsupport.util.StandardFieldMapper;
 import org.hibernate.search.integrationtest.backend.tck.testsupport.util.TckConfiguration;
-import org.hibernate.search.integrationtest.backend.tck.testsupport.util.IndexFieldStructure;
-import org.hibernate.search.integrationtest.backend.tck.testsupport.util.ValueWrapper;
 import org.hibernate.search.integrationtest.backend.tck.testsupport.util.rule.SearchSetupHelper;
-import org.hibernate.search.util.common.SearchException;
-import org.hibernate.search.util.impl.integrationtest.common.FailureReportUtils;
 import org.hibernate.search.util.impl.integrationtest.mapper.stub.StubMappingIndexManager;
 import org.hibernate.search.util.impl.integrationtest.mapper.stub.StubMappingScope;
-import org.hibernate.search.util.impl.test.SubTest;
 import org.hibernate.search.util.impl.test.annotation.TestForIssue;
 
 import org.junit.BeforeClass;
@@ -77,9 +69,6 @@ public class FieldSearchSortBaseIT<F> {
 	}
 
 	private static final String INDEX_NAME = "IndexName";
-	private static final String COMPATIBLE_INDEX_NAME = "IndexWithCompatibleFields";
-	private static final String RAW_FIELD_COMPATIBLE_INDEX_NAME = "IndexWithCompatibleRawFields";
-	private static final String INCOMPATIBLE_INDEX_NAME = "IndexWithIncompatibleFields";
 
 	private static final String DOCUMENT_1 = "1";
 	private static final String DOCUMENT_2 = "2";
@@ -89,9 +78,6 @@ public class FieldSearchSortBaseIT<F> {
 	private static final String EMPTY_2 = "empty2";
 	private static final String EMPTY_3 = "empty3";
 	private static final String EMPTY_4 = "empty4";
-
-	private static final String COMPATIBLE_INDEX_DOCUMENT_1 = "compatible_1";
-	private static final String RAW_FIELD_COMPATIBLE_INDEX_DOCUMENT_1 = "raw_field_compatible_1";
 
 	private static final int BEFORE_DOCUMENT_1_ORDINAL = 0;
 	private static final int DOCUMENT_1_ORDINAL = 1;
@@ -107,14 +93,6 @@ public class FieldSearchSortBaseIT<F> {
 	private static IndexMapping indexMapping;
 	private static StubMappingIndexManager indexManager;
 
-	private static IndexMapping compatibleIndexMapping;
-	private static StubMappingIndexManager compatibleIndexManager;
-
-	private static RawFieldCompatibleIndexMapping rawFieldCompatibleIndexMapping;
-	private static StubMappingIndexManager rawFieldCompatibleIndexManager;
-
-	private static StubMappingIndexManager incompatibleIndexManager;
-
 	@BeforeClass
 	public static void setup() {
 		setupHelper.start()
@@ -122,21 +100,6 @@ public class FieldSearchSortBaseIT<F> {
 						INDEX_NAME,
 						ctx -> indexMapping = new IndexMapping( ctx.getSchemaElement() ),
 						indexManager -> FieldSearchSortBaseIT.indexManager = indexManager
-				)
-				.withIndex(
-						COMPATIBLE_INDEX_NAME,
-						ctx -> compatibleIndexMapping = new IndexMapping( ctx.getSchemaElement() ),
-						indexManager -> compatibleIndexManager = indexManager
-				)
-				.withIndex(
-						RAW_FIELD_COMPATIBLE_INDEX_NAME,
-						ctx -> rawFieldCompatibleIndexMapping = new RawFieldCompatibleIndexMapping( ctx.getSchemaElement() ),
-						indexManager -> rawFieldCompatibleIndexManager = indexManager
-				)
-				.withIndex(
-						INCOMPATIBLE_INDEX_NAME,
-						ctx -> new IncompatibleIndexMapping( ctx.getSchemaElement() ),
-						indexManager -> incompatibleIndexManager = indexManager
 				)
 				.setup();
 
@@ -298,218 +261,6 @@ public class FieldSearchSortBaseIT<F> {
 				.hasDocRefHitsAnyOrder( INDEX_NAME, DOCUMENT_3, EMPTY_1, EMPTY_2, EMPTY_3, EMPTY_4 );
 	}
 
-	@Test
-	public void withDslConverters_dslConverterEnabled() {
-		SearchQuery<DocumentReference> query;
-		String fieldPath = getFieldWithDslConverterPath();
-
-		query = matchNonEmptyAndEmpty1Query( f -> f.field( fieldPath ).asc().missing()
-				.use( new ValueWrapper<>( getSingleValueForMissingUse( BEFORE_DOCUMENT_1_ORDINAL ) ) ) );
-		assertThat( query )
-				.hasDocRefHitsExactOrder( INDEX_NAME, EMPTY_1, DOCUMENT_1, DOCUMENT_2, DOCUMENT_3 );
-		query = matchNonEmptyAndEmpty1Query( f -> f.field( fieldPath ).asc().missing()
-				.use( new ValueWrapper<>( getSingleValueForMissingUse( BETWEEN_DOCUMENT_1_AND_2_ORDINAL ) ) ) );
-		assertThat( query )
-				.hasDocRefHitsExactOrder( INDEX_NAME, DOCUMENT_1, EMPTY_1, DOCUMENT_2, DOCUMENT_3 );
-		query = matchNonEmptyAndEmpty1Query( f -> f.field( fieldPath ).asc().missing()
-				.use( new ValueWrapper<>( getSingleValueForMissingUse( BETWEEN_DOCUMENT_2_AND_3_ORDINAL ) ) ) );
-		assertThat( query )
-				.hasDocRefHitsExactOrder( INDEX_NAME, DOCUMENT_1, DOCUMENT_2, EMPTY_1, DOCUMENT_3 );
-		query = matchNonEmptyAndEmpty1Query( f -> f.field( fieldPath ).asc().missing()
-				.use( new ValueWrapper<>( getSingleValueForMissingUse( AFTER_DOCUMENT_3_ORDINAL ) ) ) );
-		assertThat( query )
-				.hasDocRefHitsExactOrder( INDEX_NAME, DOCUMENT_1, DOCUMENT_2, DOCUMENT_3, EMPTY_1 );
-	}
-
-	@Test
-	public void withDslConverters_dslConverterDisabled() {
-		SearchQuery<DocumentReference> query;
-		String fieldPath = getFieldWithDslConverterPath();
-
-		query = matchNonEmptyAndEmpty1Query( f -> f.field( fieldPath ).asc().missing()
-				.use( getSingleValueForMissingUse( BEFORE_DOCUMENT_1_ORDINAL ), ValueConvert.NO ) );
-		assertThat( query )
-				.hasDocRefHitsExactOrder( INDEX_NAME, EMPTY_1, DOCUMENT_1, DOCUMENT_2, DOCUMENT_3 );
-		query = matchNonEmptyAndEmpty1Query( f -> f.field( fieldPath ).asc().missing()
-				.use( getSingleValueForMissingUse( BETWEEN_DOCUMENT_1_AND_2_ORDINAL ), ValueConvert.NO ) );
-		assertThat( query )
-				.hasDocRefHitsExactOrder( INDEX_NAME, DOCUMENT_1, EMPTY_1, DOCUMENT_2, DOCUMENT_3 );
-		query = matchNonEmptyAndEmpty1Query( f -> f.field( fieldPath ).asc().missing()
-				.use( getSingleValueForMissingUse( BETWEEN_DOCUMENT_2_AND_3_ORDINAL ), ValueConvert.NO ) );
-		assertThat( query )
-				.hasDocRefHitsExactOrder( INDEX_NAME, DOCUMENT_1, DOCUMENT_2, EMPTY_1, DOCUMENT_3 );
-		query = matchNonEmptyAndEmpty1Query( f -> f.field( fieldPath ).asc().missing()
-				.use( getSingleValueForMissingUse( AFTER_DOCUMENT_3_ORDINAL ), ValueConvert.NO ) );
-		assertThat( query )
-				.hasDocRefHitsExactOrder( INDEX_NAME, DOCUMENT_1, DOCUMENT_2, DOCUMENT_3, EMPTY_1 );
-	}
-
-	@Test
-	public void error_unsortable() {
-		StubMappingScope scope = indexManager.createScope();
-		String fieldPath = getNonSortableFieldPath();
-
-		SubTest.expectException( () -> {
-				scope.sort().field( fieldPath );
-		} ).assertThrown()
-				.isInstanceOf( SearchException.class )
-				.hasMessageContaining( "Sorting is not enabled for field" )
-				.hasMessageContaining( fieldPath );
-	}
-
-	@Test
-	public void error_invalidType_noDslConverter() {
-		StubMappingScope scope = indexManager.createScope();
-
-		String absoluteFieldPath = getFieldPath();
-		Object invalidValueToMatch = new InvalidType();
-
-		SubTest.expectException(
-				"field() sort with invalid parameter type for missing().use() on field " + absoluteFieldPath,
-				() -> scope.sort().field( absoluteFieldPath ).missing()
-						.use( invalidValueToMatch )
-		)
-				.assertThrown()
-				.isInstanceOf( SearchException.class )
-				.hasMessageContaining( "Unable to convert DSL parameter: " )
-				.hasMessageContaining( InvalidType.class.getName() )
-				.hasCauseInstanceOf( ClassCastException.class )
-				.satisfies( FailureReportUtils.hasContext(
-						EventContexts.fromIndexFieldAbsolutePath( absoluteFieldPath )
-				) );
-	}
-
-	@Test
-	public void error_invalidType_withDslConverter() {
-		StubMappingScope scope = indexManager.createScope();
-
-		String absoluteFieldPath = getFieldWithDslConverterPath();
-		Object invalidValueToMatch = new InvalidType();
-
-		SubTest.expectException(
-				"field() sort with invalid parameter type for missing().use() on field " + absoluteFieldPath,
-				() -> scope.sort().field( absoluteFieldPath ).missing()
-						.use( invalidValueToMatch )
-		)
-				.assertThrown()
-				.isInstanceOf( SearchException.class )
-				.hasMessageContaining( "Unable to convert DSL parameter: " )
-				.hasMessageContaining( InvalidType.class.getName() )
-				.hasCauseInstanceOf( ClassCastException.class )
-				.satisfies( FailureReportUtils.hasContext(
-						EventContexts.fromIndexFieldAbsolutePath( absoluteFieldPath )
-				) );
-	}
-
-	@Test
-	public void multiIndex_withCompatibleIndexManager_usingField() {
-		StubMappingScope scope = indexManager.createScope( compatibleIndexManager );
-
-		SearchQuery<DocumentReference> query;
-		String fieldPath = getFieldPath();
-
-		query = matchNonEmptyAndEmpty1Query( f -> f.field( fieldPath ).asc().missing()
-				.use( getSingleValueForMissingUse( BEFORE_DOCUMENT_1_ORDINAL ) ), scope );
-
-		/*
-		 * Not testing the ordering of results here because some documents have the same value.
-		 * It's not what we want to test anyway: we just want to check that fields are correctly
-		 * detected as compatible and that no exception is thrown.
-		 */
-		assertThat( query ).hasDocRefHitsAnyOrder( b -> {
-			b.doc( INDEX_NAME, EMPTY_1 );
-			b.doc( INDEX_NAME, DOCUMENT_1 );
-			b.doc( INDEX_NAME, DOCUMENT_2 );
-			b.doc( INDEX_NAME, DOCUMENT_3 );
-			b.doc( COMPATIBLE_INDEX_NAME, COMPATIBLE_INDEX_DOCUMENT_1 );
-		} );
-	}
-
-	@Test
-	public void multiIndex_withRawFieldCompatibleIndexManager_dslConverterEnabled() {
-		StubMappingScope scope = indexManager.createScope( rawFieldCompatibleIndexManager );
-
-		String fieldPath = getFieldPath();
-
-		SubTest.expectException(
-				() -> {
-					matchNonEmptyAndEmpty1Query( f -> f.field( fieldPath ).asc().missing()
-							.use( new ValueWrapper<>( getSingleValueForMissingUse( BEFORE_DOCUMENT_1_ORDINAL ) ) ), scope );
-				}
-		)
-				.assertThrown()
-				.isInstanceOf( SearchException.class )
-				.hasMessageContaining( "Multiple conflicting types to build a sort" )
-				.hasMessageContaining( "'" + fieldPath + "'" )
-				.satisfies( FailureReportUtils.hasContext(
-						EventContexts.fromIndexNames( INDEX_NAME, RAW_FIELD_COMPATIBLE_INDEX_NAME )
-				) );
-	}
-
-	@Test
-	public void multiIndex_withRawFieldCompatibleIndexManager_dslConverterDisabled() {
-		StubMappingScope scope = indexManager.createScope( rawFieldCompatibleIndexManager );
-
-		SearchQuery<DocumentReference> query;
-		String fieldPath = getFieldPath();
-
-		query = matchNonEmptyAndEmpty1Query( f -> f.field( fieldPath ).asc().missing()
-				.use( getSingleValueForMissingUse( BEFORE_DOCUMENT_1_ORDINAL ), ValueConvert.NO ), scope );
-
-		/*
-		 * Not testing the ordering of results here because some documents have the same value.
-		 * It's not what we want to test anyway: we just want to check that fields are correctly
-		 * detected as compatible and that no exception is thrown.
-		 */
-		assertThat( query ).hasDocRefHitsAnyOrder( b -> {
-			b.doc( INDEX_NAME, EMPTY_1 );
-			b.doc( INDEX_NAME, DOCUMENT_1 );
-			b.doc( INDEX_NAME, DOCUMENT_2 );
-			b.doc( INDEX_NAME, DOCUMENT_3 );
-			b.doc( RAW_FIELD_COMPATIBLE_INDEX_NAME, RAW_FIELD_COMPATIBLE_INDEX_DOCUMENT_1 );
-		} );
-	}
-
-	@Test
-	public void multiIndex_withNoCompatibleIndexManager_dslConverterEnabled() {
-		StubMappingScope scope = indexManager.createScope( incompatibleIndexManager );
-
-		String fieldPath = getFieldPath();
-
-		SubTest.expectException(
-				() -> {
-					matchNonEmptyAndEmpty1Query( f -> f.field( fieldPath ), scope );
-				}
-		)
-				.assertThrown()
-				.isInstanceOf( SearchException.class )
-				.hasMessageContaining( "Multiple conflicting types to build a sort" )
-				.hasMessageContaining( "'" + fieldPath + "'" )
-				.satisfies( FailureReportUtils.hasContext(
-						EventContexts.fromIndexNames( INDEX_NAME, INCOMPATIBLE_INDEX_NAME )
-				) );
-	}
-
-	@Test
-	public void multiIndex_withNoCompatibleIndexManager_dslConverterDisabled() {
-		StubMappingScope scope = indexManager.createScope( incompatibleIndexManager );
-
-		String fieldPath = getFieldPath();
-
-		SubTest.expectException(
-				() -> {
-					matchNonEmptyAndEmpty1Query( f -> f.field( fieldPath ), scope );
-				}
-		)
-				.assertThrown()
-				.isInstanceOf( SearchException.class )
-				.hasMessageContaining( "Multiple conflicting types to build a sort" )
-				.hasMessageContaining( "'" + fieldPath + "'" )
-				.satisfies( FailureReportUtils.hasContext(
-						EventContexts.fromIndexNames( INDEX_NAME, INCOMPATIBLE_INDEX_NAME )
-				) );
-	}
-
 	private SearchQuery<DocumentReference> matchNonEmptyQuery(
 			Function<? super SearchSortFactory, ? extends SortFinalStep> sortContributor) {
 		return matchNonEmptyQuery( sortContributor, indexManager.createScope() );
@@ -551,34 +302,19 @@ public class FieldSearchSortBaseIT<F> {
 	}
 
 	private String getFieldPath() {
-		return getAbsoluteFieldPath( objectMapping -> objectMapping.fieldModels );
-	}
-
-	private String getFieldWithDslConverterPath() {
-		return getAbsoluteFieldPath( objectMapping -> objectMapping.fieldWithDslConverterModels );
-	}
-
-	private String getNonSortableFieldPath() {
-		return getAbsoluteFieldPath( objectMapping -> objectMapping.nonSortableFieldModels );
-	}
-
-	private String getAbsoluteFieldPath(Function<AbstractObjectMapping, FieldModelsByType> fieldModels) {
 		switch ( indexFieldStructure ) {
 			case ROOT:
-				return fieldModels.apply( indexMapping ).get( fieldTypeDescriptor ).relativeFieldName;
+				return indexMapping.fieldModels.get( fieldTypeDescriptor ).relativeFieldName;
 			case IN_FLATTENED:
 				return indexMapping.flattenedObject.relativeFieldName
-						+ "." + fieldModels.apply( indexMapping.flattenedObject )
-								.get( fieldTypeDescriptor ).relativeFieldName;
+						+ "." + indexMapping.flattenedObject.fieldModels.get( fieldTypeDescriptor ).relativeFieldName;
 			case IN_NESTED:
 				return indexMapping.nestedObject.relativeFieldName
-						+ "." + fieldModels.apply( indexMapping.nestedObject )
-								.get( fieldTypeDescriptor ).relativeFieldName;
+						+ "." + indexMapping.nestedObject.fieldModels.get( fieldTypeDescriptor ).relativeFieldName;
 			case IN_NESTED_TWICE:
 				return indexMapping.nestedObject.relativeFieldName
 						+ "." + indexMapping.nestedObject.nestedObject.relativeFieldName
-						+ "." + fieldModels.apply( indexMapping.nestedObject.nestedObject )
-								.get( fieldTypeDescriptor ).relativeFieldName;
+						+ "." + indexMapping.nestedObject.nestedObject.fieldModels.get( fieldTypeDescriptor ).relativeFieldName;
 			default:
 				throw new IllegalStateException( "Unexpected value: " + indexFieldStructure );
 		}
@@ -587,7 +323,6 @@ public class FieldSearchSortBaseIT<F> {
 	private static void initDocument(IndexMapping indexMapping, DocumentElement document, Integer ordinal) {
 		forEachSupportedTypeDescriptor( typeDescriptor -> {
 			addValue( document, indexMapping.fieldModels, typeDescriptor, ordinal );
-			addValue( document, indexMapping.fieldWithDslConverterModels, typeDescriptor, ordinal );
 		} );
 
 		// Note: these objects must be single-valued for these tests
@@ -597,11 +332,8 @@ public class FieldSearchSortBaseIT<F> {
 
 		forEachSupportedTypeDescriptor( typeDescriptor -> {
 			addValue( flattenedObject, indexMapping.flattenedObject.fieldModels, typeDescriptor, ordinal );
-			addValue( flattenedObject, indexMapping.flattenedObject.fieldWithDslConverterModels, typeDescriptor, ordinal );
 			addValue( nestedObject, indexMapping.nestedObject.fieldModels, typeDescriptor, ordinal );
-			addValue( nestedObject, indexMapping.nestedObject.fieldWithDslConverterModels, typeDescriptor, ordinal );
 			addValue( nestedObjectInNestedObject, indexMapping.nestedObject.nestedObject.fieldModels, typeDescriptor, ordinal );
-			addValue( nestedObjectInNestedObject, indexMapping.nestedObject.nestedObject.fieldWithDslConverterModels, typeDescriptor, ordinal );
 		} );
 	}
 
@@ -649,30 +381,12 @@ public class FieldSearchSortBaseIT<F> {
 				document -> initDocument( indexMapping, document, null ) );
 		plan.execute().join();
 
-		plan = compatibleIndexManager.createIndexingPlan();
-		plan.add( referenceProvider( COMPATIBLE_INDEX_DOCUMENT_1 ),
-				document -> initDocument( compatibleIndexMapping, document, DOCUMENT_1_ORDINAL ) );
-		plan.execute().join();
-
-		plan = rawFieldCompatibleIndexManager.createIndexingPlan();
-		plan.add( referenceProvider( RAW_FIELD_COMPATIBLE_INDEX_DOCUMENT_1 ),
-				document -> initDocument( rawFieldCompatibleIndexMapping, document, DOCUMENT_1_ORDINAL ) );
-		plan.execute().join();
-
 		// Check that all documents are searchable
 		SearchQuery<DocumentReference> query = indexManager.createScope().query()
 				.where( f -> f.matchAll() )
 				.toQuery();
 		assertThat( query )
 				.hasDocRefHitsAnyOrder( INDEX_NAME, DOCUMENT_1, DOCUMENT_2, DOCUMENT_3, EMPTY_1, EMPTY_2, EMPTY_3, EMPTY_4 );
-		query = compatibleIndexManager.createScope().query()
-				.where( f -> f.matchAll() )
-				.toQuery();
-		assertThat( query ).hasDocRefHitsAnyOrder( COMPATIBLE_INDEX_NAME, COMPATIBLE_INDEX_DOCUMENT_1 );
-		query = rawFieldCompatibleIndexManager.createScope().query()
-				.where( f -> f.matchAll() )
-				.toQuery();
-		assertThat( query ).hasDocRefHitsAnyOrder( RAW_FIELD_COMPATIBLE_INDEX_NAME, RAW_FIELD_COMPATIBLE_INDEX_DOCUMENT_1 );
 	}
 
 	private static void forEachSupportedTypeDescriptor(Consumer<FieldTypeDescriptor<?>> action) {
@@ -683,22 +397,10 @@ public class FieldSearchSortBaseIT<F> {
 
 	private static class AbstractObjectMapping {
 		final FieldModelsByType fieldModels;
-		final FieldModelsByType fieldWithDslConverterModels;
-		final FieldModelsByType nonSortableFieldModels;
 
 		AbstractObjectMapping(IndexSchemaElement self,
 				Consumer<StandardIndexFieldTypeOptionsStep<?, ?>> additionalConfiguration) {
 			fieldModels = FieldModelsByType.mapSupported( self, "", additionalConfiguration );
-			fieldWithDslConverterModels = FieldModelsByType.mapSupported(
-					self, "converted_",
-					additionalConfiguration.andThen(
-							c -> c.dslConverter( ValueWrapper.class, ValueWrapper.toIndexFieldConverter() )
-					)
-			);
-			nonSortableFieldModels = FieldModelsByType.mapSupported(
-					self, "nonSortable_",
-					additionalConfiguration.andThen( c -> c.sortable( Sortable.NO ) )
-			);
 		}
 	}
 
@@ -764,54 +466,6 @@ public class FieldSearchSortBaseIT<F> {
 		}
 	}
 
-	private static class RawFieldCompatibleIndexMapping extends IndexMapping {
-		RawFieldCompatibleIndexMapping(IndexSchemaElement root) {
-			/*
-			 * Add fields with the same name as the fieldModels from IndexMapping,
-			 * but with an incompatible DSL converter.
-			 */
-			super( root, c -> c.dslConverter( ValueWrapper.class, ValueWrapper.toIndexFieldConverter() ) );
-		}
-	}
-
-	private static class IncompatibleIndexMapping {
-		IncompatibleIndexMapping(IndexSchemaElement root) {
-			/*
-			 * Add fields with the same name as the supportedFieldModels from IndexMapping,
-			 * but with an incompatible type.
-			 */
-			mapFieldsWithIncompatibleType( root );
-
-			IndexSchemaObjectField nestedObject =
-					root.objectField( "nestedObject", ObjectFieldStorage.NESTED );
-			nestedObject.toReference();
-			mapFieldsWithIncompatibleType( nestedObject );
-
-			IndexSchemaObjectField flattenedObject =
-					root.objectField( "flattenedObject", ObjectFieldStorage.FLATTENED );
-			flattenedObject.toReference();
-			mapFieldsWithIncompatibleType( flattenedObject );
-
-			IndexSchemaObjectField nestedObjectInNestedObject =
-					nestedObject.objectField( "nestedObject", ObjectFieldStorage.NESTED );
-			nestedObjectInNestedObject.toReference();
-			mapFieldsWithIncompatibleType( nestedObjectInNestedObject );
-		}
-
-		private static void mapFieldsWithIncompatibleType(IndexSchemaElement parent) {
-			forEachSupportedTypeDescriptor( typeDescriptor -> {
-				StandardFieldMapper<?, IncompatibleFieldModel> mapper;
-				if ( Integer.class.equals( typeDescriptor.getJavaType() ) ) {
-					mapper = IncompatibleFieldModel.mapper( context -> context.asLong() );
-				}
-				else {
-					mapper = IncompatibleFieldModel.mapper( context -> context.asInteger() );
-				}
-				mapper.map( parent, "" + typeDescriptor.getUniqueName() );
-			} );
-		}
-	}
-
 	private static class FieldModelsByType {
 		public static FieldModelsByType mapSupported(IndexSchemaElement parent, String prefix,
 				Consumer<StandardIndexFieldTypeOptionsStep<?, ?>> additionalConfiguration) {
@@ -848,22 +502,6 @@ public class FieldSearchSortBaseIT<F> {
 
 		private FieldModel(IndexFieldReference<F> reference, String relativeFieldName) {
 			this.reference = reference;
-			this.relativeFieldName = relativeFieldName;
-		}
-	}
-
-	private static class IncompatibleFieldModel {
-		static <F> StandardFieldMapper<F, IncompatibleFieldModel> mapper(
-				Function<IndexFieldTypeFactory, StandardIndexFieldTypeOptionsStep<?, F>> configuration) {
-			return StandardFieldMapper.of(
-					configuration,
-					(reference, name) -> new IncompatibleFieldModel( name )
-			);
-		}
-
-		final String relativeFieldName;
-
-		private IncompatibleFieldModel(String relativeFieldName) {
 			this.relativeFieldName = relativeFieldName;
 		}
 	}
